src: db: introduce db model and migration script
Introduces the file that builds the model to be used by kw to store all
data related to statistics options, pomodoro sessions and config files.
Adds a script to handle the data migration from the files in
`KW_DATA_DIR` to the newly introduced database.

Thanks: Magali Lemes <magalilemes00@gmail.com> for the help modelling
  the database.
Signed-off-by: Rubens Gomes Neto <rubens.gomes.neto@alumni.usp.br>